+++ test/tx-data/declare-candidacy.test.js
@@ -6,7 +6,7 @@ describe('MinterTxDataDeclareCandidacy', () => {
     test('rlp encoded fields', () => {
         const serializedTxData = (new MinterTxDataDeclareCandidacy({
             address: toBuffer('Mx7633980c000139dd3bd24a3f54e06474fa941e16'),
-            pubKey: toBuffer('Mpf9e036839a29f7fba2d5394bd489eda927ccb95acc99e506e688e4888082b3a3'),
+            publicKey: toBuffer('Mpf9e036839a29f7fba2d5394bd489eda927ccb95acc99e506e688e4888082b3a3'),
             commission: 10,
             coin: coinToBuffer('MNT'),
             stake: 1000,