+++ test/tx-data/set-candidate-off.test.js
@@ -5,7 +5,7 @@ import decodeToArray from '../decode-to-array';
 describe('MinterTxDataSetCandidateOff', () => {
     test('rlp encoded fields', () => {
         const serializedTxData = (new MinterTxDataSetCandidateOff({
-            pubKey: toBuffer('Mpf9e036839a29f7fba2d5394bd489eda927ccb95acc99e506e688e4888082b3a3'),
+            publicKey: toBuffer('Mpf9e036839a29f7fba2d5394bd489eda927ccb95acc99e506e688e4888082b3a3'),
         })).serialize();
 
         expect(decodeToArray(serializedTxData))